Moved the Results class into inference.py
